slides: updated slides with 'syno' command
Also, decided on naming convention for .tpp files. Whenever there exists
a man page for a specific git topic, the corresponding .tpp file should
be called the same as the man page, eg git-config or githooks. The titl
of a page should be the git man page name, with section number, eg
git-config(1). Any extra slides on a topic should be named
<original>-<extra>.tpp, eg git-config-alias.tpp, and have a titl
<original> --- <extra>, eg git-config(1) --- alias.

If not man page exists, any .tpp file name and titl will do.